+++ src/AntaresProto.js
@@ -1,0 +1,2 @@
+// properties Antares has that can be set during construction - becomes prototype
+export const AntaresProto = {}