+++ src/config.js
@@ -1,3 +1,5 @@
+import { AntaresProto } from './AntaresProto'
+
 // Configs that can be overridden by AntaresInit
 export const Agents = {
     server: () => ((typeof process !== 'undefined') && process.env &&
@@ -12,6 +14,8 @@ export const Types = {}
 export const ReducerForKey = []
 export const ViewReducer = []
 export const MetaEnhancers = [
-    () => ({ actionId: Math.floor(Math.random() * 10000) })
+    // like airplane flight numbers, only guaranteed unique within small time window
+    () => ({ actionId: Math.floor(Math.random() * 10000) }),
+    () => ({ originAgentId: AntaresProto.agentId })
 ]
 export const DispatchProxy = []