+++ test/unit/factories.js
@@ -1,5 +1,6 @@
 export const minimalConfig = {
     client: () => true,
     defineDispatchEndpoint: () => null,
-    defineRemoteActionsProducer: () => null
+    defineRemoteActionsProducer: () => null,
+    newId: () => Math.floor(Math.random() * 10000)
 }